Backend/app/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.1.0
beautifulsoup4==4.13.3
billiard==4.2.1
bs4==0.0.2
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
cssselect==1.2.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
elastic-transport==8.17.0
elasticsearch==8.17.1
fastapi==0.115.7
filelock==3.17.0
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain
langchain-community==0.3.16
langchain-core==0.3.33
langchain-groq==0.2.4
langchain-text-splitters==0.3.5
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.3.6
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.26.1
minio==7.2.15
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy
orjson==3.10.15
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pdfquery==0.4.3
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.2.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.11
PyMuPDF==1.25.2
pymupdf4llm==0.0.17
pypdfium2==4.30.1
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
roman==5.0
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlite-vec==0.1.6
starlette==0.45.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
zstandard==0.23.0

